pyproject.toml
[project]
name = "video-dubber"
version = "0.1.0"
description = "Async pipeline for dubbing videos with GPT-4o voices."
authors = [{ name = "MGasc" }]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.42.0",
    "moviepy>=1.0.3",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.1",
    "audioop-lts>=0.2.2",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
